move kv to pkg
