+++ server/region_syncer/history_buffer.go
@@ -19,9 +19,9 @@ import (
 
 	"github.com/pingcap/log"
 	"github.com/tikv/pd/pkg/errs"
+	"github.com/tikv/pd/pkg/storage/kv"
 	"github.com/tikv/pd/pkg/utils/syncutil"
 	"github.com/tikv/pd/server/core"
-	"github.com/tikv/pd/server/storage/kv"
 	"go.uber.org/zap"
 )
