+++ server/region_syncer/history_buffer_test.go
@@ -19,8 +19,8 @@ import (
 
 	"github.com/pingcap/kvproto/pkg/metapb"
 	"github.com/stretchr/testify/require"
+	"github.com/tikv/pd/pkg/storage/kv"
 	"github.com/tikv/pd/server/core"
-	"github.com/tikv/pd/server/storage/kv"
 )
 
 func TestBufferSize(t *testing.T) {
